remove ::int Postgres conversions, since Resql does it.

+++ DSL/Resql/efti/POST/get_users.sql
@@ -17,4 +17,4 @@ SELECT * FROM (
   ORDER BY id, created_at DESC
 ) latest
 WHERE is_user_active = true
-LIMIT COALESCE(:limit::int, 20) OFFSET COALESCE(:offset::int, 0);
+LIMIT COALESCE(:limit, 20) OFFSET COALESCE(:offset, 0);
